+++ apps/backoffice-v2/src/common/components/molecules/ProcessTracker/constants.tsx
@@ -90,5 +90,13 @@ export const tagToIcon = {
   [StateTag.REVISION]: Icon.REFRESH,
 } as const;
 
-export const pluginsWhiteList = ['kyb', 'ubo', 'company_sanctions', 'merchant_monitoring'] as const;
+export const pluginsWhiteList = [
+  'kyb',
+  'ubo',
+  'company_sanctions',
+  'merchant_monitoring',
+  'businessInformation',
+  'companySanctions',
+  'merchantMonitoring',
+] as const;
 export const DEFAULT_PROCESS_TRACKER_PROCESSES = ['collection-flow', 'third-party', 'ubos'];