+++ apps/backoffice-v2/src/pages/MerchantMonitoring/hooks/useMerchantMonitoringLogic/useMerchantMonitoringLogic.tsx
@@ -1,31 +1,52 @@
 import { useBusinessReportsQuery } from '@/domains/business-reports/hooks/queries/useBusinessReportsQuery/useBusinessReportsQuery';
 import { useZodSearchParams } from '@/common/hooks/useZodSearchParams/useZodSearchParams';
 import { getMerchantMonitoringSearchSchema } from '@/pages/MerchantMonitoring/get-merchant-monitoring-search-schema';
-import { useAuthenticatedUserQuery } from '@/domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
 import { usePagination } from '@/common/hooks/usePagination/usePagination';
 import { useLocale } from '@/common/hooks/useLocale/useLocale';
 import { useCustomerQuery } from '@/domains/customer/hook/queries/useCustomerQuery/useCustomerQuery';
+import { useSearch } from '@/common/hooks/useSearch/useSearch';
 
 export const useMerchantMonitoringLogic = () => {
   const locale = useLocale();
   const { data: customer } = useCustomerQuery();
-  const { data: session } = useAuthenticatedUserQuery();
-  const MerchantMonitoringSearchSchema = getMerchantMonitoringSearchSchema(session?.user?.id);
-  const [{ page, pageSize, sortBy, sortDir }] = useZodSearchParams(MerchantMonitoringSearchSchema);
-  const { data: businessReports, isLoading: isLoadingBusinessReports } = useBusinessReportsQuery({
+
+  const MerchantMonitoringSearchSchema = getMerchantMonitoringSearchSchema();
+
+  const [{ page, pageSize, sortBy, sortDir, search: searchParamValue }] = useZodSearchParams(
+    MerchantMonitoringSearchSchema,
+  );
+
+  const { search: searchTerm, onSearch } = useSearch({
+    initialSearch: searchParamValue,
+  });
+
+  const search = searchTerm as string;
+
+  const {
+    data = { businessReports: [], meta: { totalItems: 0, totalPages: 0 } },
+    isLoading: isLoadingBusinessReports,
+  } = useBusinessReportsQuery({
     reportType: 'MERCHANT_REPORT_T1',
+    search,
     page,
     pageSize,
     sortBy,
     sortDir,
   });
+
   const { onPaginate, onPrevPage, onNextPage } = usePagination();
-  const isLastPage = (businessReports?.length ?? 0) < pageSize || businessReports?.length === 0;
+
+  const isLastPage =
+    (data?.businessReports?.length ?? 0) < pageSize || data?.businessReports?.length === 0;
 
   return {
+    totalPages: data?.meta.totalPages || 0,
+    totalItems: data?.meta.totalItems || 0,
     hideCreateMerchantMonitoringButton: customer?.config?.hideCreateMerchantMonitoringButton,
-    businessReports,
+    businessReports: data?.businessReports || [],
     isLoadingBusinessReports,
+    search,
+    onSearch,
     page,
     onPrevPage,
     onNextPage,