+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ballerine/ui
 
+## 0.5.10
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.15
+
+## 0.5.9
+
+### Patch Changes
+
+- Fixed dropdown input text overflow
+
 ## 0.5.8
 
 ### Patch Changes